use long long

+++ basis.hpp
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <vector>
-#include <cstdint>
 #include <complex>
 
 class MatrixAccessor;
@@ -11,16 +10,16 @@ class CellComm;
 
 class WellSeparatedApproximation {
 private:
-  int64_t lbegin;
-  int64_t lend;
+  long long lbegin;
+  long long lend;
   std::vector<std::vector<double>> M;
 
 public:
   WellSeparatedApproximation() : lbegin(0), lend(0) {}
-  WellSeparatedApproximation(const MatrixAccessor& eval, double epi, int64_t rank, int64_t lbegin, int64_t lend, const Cell cells[], const CSR& Far, const double bodies[], const WellSeparatedApproximation& upper);
+  WellSeparatedApproximation(const MatrixAccessor& eval, double epi, long long rank, long long lbegin, long long lend, const Cell cells[], const CSR& Far, const double bodies[], const WellSeparatedApproximation& upper);
 
-  int64_t fbodies_size_at_i(int64_t i) const;
-  const double* fbodies_at_i(int64_t i) const;
+  long long fbodies_size_at_i(long long i) const;
+  const double* fbodies_at_i(long long i) const;
 };
 
 class ClusterBasis {
@@ -31,19 +30,19 @@ private:
 
   std::vector<double> Sdata;
   std::vector<const double*> S;
-  std::vector<int64_t> elementsOnRow;
-  std::vector<int64_t> localChildOffsets;
-  std::vector<int64_t> localChildLrDims;
-  int64_t localChildIndex;
+  std::vector<long long> elementsOnRow;
+  std::vector<long long> localChildOffsets;
+  std::vector<long long> localChildLrDims;
+  long long localChildIndex;
 
 public:
-  std::vector<int64_t> Dims;
-  std::vector<int64_t> DimsLr;
+  std::vector<long long> Dims;
+  std::vector<long long> DimsLr;
   std::vector<const std::complex<double>*> Q;
   std::vector<std::complex<double>*> R;
 
-  std::vector<int64_t> CRows;
-  std::vector<int64_t> CCols;
+  std::vector<long long> CRows;
+  std::vector<long long> CCols;
   std::vector<const std::complex<double>*> C;
   
   ClusterBasis() {}
@@ -62,10 +61,10 @@ private:
   const Cell* Cells;
   const CSR* Near;
   const CellComm* Comm;
-  int64_t Levels;
+  long long Levels;
 
 public:
-  MatVec(const MatrixAccessor& eval, const ClusterBasis basis[], const double bodies[], const Cell cells[], const CSR& near, const CellComm comm[], int64_t levels);
+  MatVec(const MatrixAccessor& eval, const ClusterBasis basis[], const double bodies[], const Cell cells[], const CSR& near, const CellComm comm[], long long levels);
 
-  void operator() (int64_t nrhs, std::complex<double> X[], int64_t ldX) const;
+  void operator() (long long nrhs, std::complex<double> X[], long long ldX) const;
 };
