+++ build_tree.cpp
@@ -8,14 +8,14 @@
 #include <array>
 #include <set>
 
-void get_bounds(const double* bodies, int64_t nbodies, double R[], double C[]) {
+void get_bounds(const double* bodies, long long nbodies, double R[], double C[]) {
   double Xmin[3];
   double Xmax[3];
   Xmin[0] = Xmax[0] = bodies[0];
   Xmin[1] = Xmax[1] = bodies[1];
   Xmin[2] = Xmax[2] = bodies[2];
 
-  for (int64_t i = 1; i < nbodies; i++) {
+  for (long long i = 1; i < nbodies; i++) {
     const double* x_bi = &bodies[i * 3];
     Xmin[0] = fmin(x_bi[0], Xmin[0]);
     Xmin[1] = fmin(x_bi[1], Xmin[1]);
@@ -39,7 +39,7 @@ void get_bounds(const double* bodies, int64_t nbodies, double R[], double C[]) {
   R[2] = (d2 == 0. && Xmin[2] == 0.) ? 0. : (1.e-8 + d2 / 2.);
 }
 
-void sort_bodies(double* bodies, int64_t nbodies, int64_t sdim) {
+void sort_bodies(double* bodies, long long nbodies, long long sdim) {
   std::array<double, 3>* bodies3 = reinterpret_cast<std::array<double, 3>*>(bodies);
   std::array<double, 3>* bodies3_end = reinterpret_cast<std::array<double, 3>*>(&bodies[3 * nbodies]);
   std::sort(bodies3, bodies3_end, [=](std::array<double, 3>& i, std::array<double, 3>& j)->bool {
@@ -74,31 +74,31 @@ int admis_check(double theta, const double C1[], const double C2[], const double
   return (int)(dC > dR);
 }
 
-void buildTree(Cell* cells, double* bodies, int64_t nbodies, int64_t levels) {
+void buildTree(Cell* cells, double* bodies, long long nbodies, long long levels) {
   Cell* root = &cells[0];
   root->Body[0] = 0;
   root->Body[1] = nbodies;
   get_bounds(bodies, nbodies, root->R, root->C);
 
-  int64_t nleaf = (int64_t)1 << levels;
-  int64_t len = 1;
-  for (int64_t i = 0; i < nleaf - 1; i++) {
+  long long nleaf = (long long)1 << levels;
+  long long len = 1;
+  for (long long i = 0; i < nleaf - 1; i++) {
     Cell* ci = &cells[i];
     ci->Child[0] = -1;
     ci->Child[1] = -1;
 
-    int64_t sdim = 0;
+    long long sdim = 0;
     double maxR = ci->R[0];
     if (ci->R[1] > maxR)
     { sdim = 1; maxR = ci->R[1]; }
     if (ci->R[2] > maxR)
     { sdim = 2; maxR = ci->R[2]; }
 
-    int64_t i_begin = ci->Body[0];
-    int64_t i_end = ci->Body[1];
-    int64_t nbody_i = i_end - i_begin;
+    long long i_begin = ci->Body[0];
+    long long i_end = ci->Body[1];
+    long long nbody_i = i_end - i_begin;
     sort_bodies(&bodies[i_begin * 3], nbody_i, sdim);
-    int64_t loc = i_begin + nbody_i / 2;
+    long long loc = i_begin + nbody_i / 2;
 
     Cell* c0 = &cells[len];
     Cell* c1 = &cells[len + 1];
@@ -116,11 +116,11 @@ void buildTree(Cell* cells, double* bodies, int64_t nbodies, int64_t levels) {
   }
 }
 
-void buildTreeBuckets(Cell* cells, const double* bodies, const int64_t buckets[], int64_t levels) {
-  int64_t nleaf = (int64_t)1 << levels;
-  int64_t count = 0;
-  for (int64_t i = 0; i < nleaf; i++) {
-    int64_t ci = i + nleaf - 1;
+void buildTreeBuckets(Cell* cells, const double* bodies, const long long buckets[], long long levels) {
+  long long nleaf = (long long)1 << levels;
+  long long count = 0;
+  for (long long i = 0; i < nleaf; i++) {
+    long long ci = i + nleaf - 1;
     cells[ci].Child[0] = -1;
     cells[ci].Child[1] = -1;
     cells[ci].Body[0] = count;
@@ -129,11 +129,11 @@ void buildTreeBuckets(Cell* cells, const double* bodies, const int64_t buckets[]
     count = count + buckets[i];
   }
 
-  for (int64_t i = nleaf - 2; i >= 0; i--) {
-    int64_t c0 = (i << 1) + 1;
-    int64_t c1 = (i << 1) + 2;
-    int64_t begin = cells[c0].Body[0];
-    int64_t len = cells[c1].Body[1] - begin;
+  for (long long i = nleaf - 2; i >= 0; i--) {
+    long long c0 = (i << 1) + 1;
+    long long c1 = (i << 1) + 2;
+    long long begin = cells[c0].Body[0];
+    long long len = cells[c1].Body[1] - begin;
     cells[i].Child[0] = c0;
     cells[i].Child[1] = c0 + 2;
     cells[i].Body[0] = begin;
@@ -142,7 +142,7 @@ void buildTreeBuckets(Cell* cells, const double* bodies, const int64_t buckets[]
   }
 }
 
-void getList(char NoF, std::vector<std::pair<int64_t, int64_t>>& rels, const Cell cells[], int64_t i, int64_t j, double theta) {
+void getList(char NoF, std::vector<std::pair<long long, long long>>& rels, const Cell cells[], long long i, long long j, double theta) {
   int admis = admis_check(theta, cells[i].C, cells[j].C, cells[i].R, cells[j].R);
   int write_far = NoF == 'F' || NoF == 'f';
   int write_near = NoF == 'N' || NoF == 'n';
@@ -150,8 +150,8 @@ void getList(char NoF, std::vector<std::pair<int64_t, int64_t>>& rels, const Cel
     rels.emplace_back(i, j);
   
   if (!admis && cells[i].Child[0] >= 0 && cells[j].Child[0] >= 0)
-    for (int64_t k = cells[i].Child[0]; k < cells[i].Child[1]; k++)
-      for (int64_t l = cells[j].Child[0]; l < cells[j].Child[1]; l++)
+    for (long long k = cells[i].Child[0]; k < cells[i].Child[1]; k++)
+      for (long long l = cells[j].Child[0]; l < cells[j].Child[1]; l++)
         getList(NoF, rels, cells, k, l, theta);
 }
 
@@ -162,40 +162,40 @@ Cell::Cell() {
   R[0] = R[1] = R[2] = 0.;
 }
 
-CSR::CSR(char NoF, int64_t ncells, const Cell* cells, double theta) {
-  std::vector<std::pair<int64_t, int64_t>> LIL;
+CSR::CSR(char NoF, long long ncells, const Cell* cells, double theta) {
+  std::vector<std::pair<long long, long long>> LIL;
   getList(NoF, LIL, cells, 0, 0, theta);
   std::sort(LIL.begin(), LIL.end());
 
-  int64_t len = LIL.size();
+  long long len = LIL.size();
   RowIndex.resize(ncells + 1);
   ColIndex.resize(len);
   std::transform(LIL.begin(), LIL.end(), ColIndex.begin(), 
-    [](const std::pair<int64_t, int64_t>& i) { return i.second; });
+    [](const std::pair<long long, long long>& i) { return i.second; });
 
   RowIndex[0] = 0;
-  for (int64_t n = 1; n <= ncells; n++)
+  for (long long n = 1; n <= ncells; n++)
     RowIndex[n] = std::distance(LIL.begin(), 
       std::find_if(LIL.begin() + RowIndex[n - 1], LIL.end(), 
-        [=](const std::pair<int64_t, int64_t>& i) { return n <= i.first; }));
+        [=](const std::pair<long long, long long>& i) { return n <= i.first; }));
 }
 
 CSR::CSR(const CSR& A, const CSR& B) {
-  int64_t M = A.RowIndex.size() - 1;
+  long long M = A.RowIndex.size() - 1;
   RowIndex.resize(M + 1);
   ColIndex.clear();
   RowIndex[0] = 0;
 
-  for (int64_t y = 0; y < M; y++) {
-    std::set<int64_t> cols;
-    std::for_each(&A.ColIndex[A.RowIndex[y]], &A.ColIndex[A.RowIndex[y + 1]], [&](int64_t x) {
-      std::for_each(&B.ColIndex[B.RowIndex[x]], &B.ColIndex[B.RowIndex[x + 1]], [&](int64_t z) {
+  for (long long y = 0; y < M; y++) {
+    std::set<long long> cols;
+    std::for_each(&A.ColIndex[A.RowIndex[y]], &A.ColIndex[A.RowIndex[y + 1]], [&](long long x) {
+      std::for_each(&B.ColIndex[B.RowIndex[x]], &B.ColIndex[B.RowIndex[x + 1]], [&](long long z) {
         cols.insert(z);
       });
     });
 
-    for (std::set<int64_t>::iterator i = cols.begin(); i != cols.end(); i = std::next(i))
+    for (std::set<long long>::iterator i = cols.begin(); i != cols.end(); i = std::next(i))
       ColIndex.push_back(*i);
-    RowIndex[y + 1] = (int64_t)ColIndex.size();
+    RowIndex[y + 1] = (long long)ColIndex.size();
   }
 }
